% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{APA}
\alias{APA}
\title{APA() compute Area Potentially Available (APA)}
\usage{
APA(x, y, dbh, plot_x, plot_y, weighted = TRUE)
}
\arguments{
\item{x}{: double | vector of x coordinates for each tree}

\item{y}{: double | vector of y coordinates for each tree}

\item{dbh}{: double | vector of dbh for each tree}

\item{plot_x}{: double | vector of plot corners' x coordinates (bottom left, top right)}

\item{plot_y}{: double | vector of plot corners' y coordinates (bottom left, top right)}

\item{weighted}{: bool   | TRUE = use dbh^2 as weighting factor for polygon construction, FALSE = unweighted (default)}
}
\value{
Returns a vector of APA values (in square units of measure used for the coordinates) for each tree in their original order.
}
\description{
Computes Area Potentially Available (APA) using Brown's (1965) method. Polygons are constructed 
around the subject tree by the intersection of the perpendicular bisectors of the distance between 
the subject tree and competitors (creating a Voronoi tesselation of the plot).

NOTE: the \code{weighted} option currently does not work.
}
\examples{
data(treelistxy)
min_x <- min(treelistxy$x)
min_y <- min(treelistxy$y)
max_x <- max(treelistxy$x)
max_y <- max(treelistxy$y)
poly_x <- c(min_x, max_x)
poly_y <- c(min_y, max_y)
APA( treelistxy$x, treelistxy$y, treelistxy$dbh, poly_x, poly_y, F )

}
