% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Arney_CSI}
\alias{Arney_CSI}
\title{Arney_CSI() compute Arney's competitive stress index.}
\usage{
Arney_CSI(x, y, dbh, mcw)
}
\arguments{
\item{x}{: double | vector of x coordinates of trees on plot (in same units as mcw)}

\item{y}{: double | vector of y coordinates of trees on plot (in same units as mcw)}

\item{dbh}{: double | vector of diameter at breast height}

\item{mcw}{: double | vector of maximum crown width for the tree record}

\item{imperial_units}{: bool   | TRUE = imperial, FALSE = metric (default)}
}
\value{
Returns the Arney's CSI for each tree.
}
\description{
Compute Arney's Competitive Stress Index as described in Arney (1973). CSI is the sum of percentage competing trees crown area overlaping
a subject tree to the subject tree's crown area.

\eqn{CSI_i = 100 \sum{\frac{AO_j}{CA_i}}}

where \eqn{AO_j} is the area of overlap of tree j on subject tree i, \eqn{CA_i} is the crown area of the subject tree i, and \eqn{CSI_i} is the
competitive stress index for tree i.

This version currently does not adjust for edge effects.
}
\examples{
data(treelistxy)
Arney_CSI( treelistxy$x, treelistxy$y, treelistxy$dbh, mcw( treelistxy$fia, treelistxy$dbh ) )

}
