% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ba}
\alias{ba}
\title{ba() compute basal area per unit area.}
\usage{
ba(dbh, expansion, imperial_units = TRUE)
}
\arguments{
\item{dbh}{: double | vector of diameter at breast height}

\item{expansion}{: double | vector of expansion factors}

\item{imperial_units}{: bool   | TRUE = imperial (default), FALSE = metric}
}
\value{
Returns the basal area per unit area or NAN if there are improper arguments.
}
\description{
Computes the basal area per unit area of the trees in the tree list. 

\eqn{ba = \sum{dbh_i^2 expansion_i k}}

where k converts squared diameters to square feet or meters depending on \code{imperial_units}.
}
\examples{
data(treelist)
ba( treelist$dbh, treelist$tpa )

}
