% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bal}
\alias{bal}
\title{bal() compute basal area in larger trees.}
\usage{
bal(dbh, expansion, imperial_units = TRUE)
}
\arguments{
\item{dbh}{: double | vector of diameter at breast height}

\item{expansion}{: double | vector of expansion factors}

\item{imperial_units}{: bool   | TRUE = imperial (default), FALSE = metric}
}
\value{
Returns numeric vector with the basal area in larger trees in the original tree order (unsorted):
}
\description{
Sorts the dbh and expansion vectors in decreasing order of dbh and computes cumulative basal area in larger trees. Ties in dbh are
handled appropriately (each tree has the same bal).
}
\examples{
data(treelist)
bal( treelist$dbh, treelist$tpa )

}
