% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ccf}
\alias{ccf}
\title{ccf() compute crown competition factor}
\usage{
ccf(crown_width, expansion, imperial_units = TRUE)
}
\arguments{
\item{crown_width}{: double | vector of open-grown crown widths}

\item{expansion}{: double | vector of expansion factors}

\item{imperial_units}{: bool   | TRUE = imperial (default), FALSE = metric}
}
\value{
Returns the crown competition factor.
}
\description{
Compute Crown Competition Factor (Krajicek, et al. 1961). 

Crown competition factor is the ratio of the open-grown crown area of all trees as a percentage of an acre:
\eqn{CCF = 100 \sum{CA expansion} /43560}, 
where CA is the open-grown crown area of a given tree.

The function allows for computing the equivalent index in metric units.
}
\examples{
data(treelist)
library( dplyr )

# use Hann's (1998) largest crown width equations Douglas-fir, western hemlock, bigleaf maple, and red alder
lcw.parms <- data.frame( species=c(202,263,312,351), 
                         b0=c(0.0,         0.0, 0.0,      0.3227140 ),
                         b1=c(0.004363240, 0.0, 0.0,      0.0       ),
                         b2=c(0.6020020,   0.0, 1.470180, 0.0       ))

temp <- treelist \%>\% mutate( mcw.hat=mcw( species, dbh ) ) \%>\%
            inner_join( lcw.parms, by="species" ) \%>\%
            mutate( crown_length = height - htlc,
                    cr = crown_length/height,
                    lcw = mcw.hat * cr^(b0+b1*crown_length+b2*(dbh/height)) )

ccf( temp$lcw, temp$tpa )

}
