% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cch}
\alias{cch}
\title{cch() compute closure at tree tip without using interpolation.}
\usage{
cch(
  species,
  dbh,
  height,
  crown_length,
  dacb,
  lcw,
  expansion,
  parameters,
  imperial_units = TRUE
)
}
\arguments{
\item{species}{: int    | vector of species codes}

\item{dbh}{: double | vector of diameter at breast height}

\item{height}{: double | vector of total height}

\item{crown_length}{: double | vector of crown lengths}

\item{dacb}{: double | vector of distance above crown base to largest width for each tree}

\item{lcw}{: double | vector of largest crown width for each tree}

\item{expansion}{: double | vector of expansion factors}

\item{parameters}{: data.frame | data.frame of parameters by species (see below)}

\item{imperial_units}{: bool   | TRUE = imperial (default), FALSE = metric}
}
\value{
A vector of \code{cch} values for each tree.
}
\description{
Computes the crown area as a fraction of an acre or hectare (depending on \code{imperial_units}) that lies in a horizontal
plane tangential to the tip of the tree.

cch() depends on \code{dacb} and \code{lcw} which must be estimated for each tree independently (normally species and size dependent).

The \code{parameters} \code{data.frame} contains species-specific coefficients for an equation that adjusts \code{lcw} for the position within the crown of the form:

\eqn{adjustment = rp^{(\beta_0 + \beta_1 rp^{0.5} + \beta_2 height / dbh)}}

where \code{rp} is the relative position in the crown, and \code{adjustment} is the estimated percentage of the \code{lcw} at that point
in the crown. The \code{parameters} are the \eqn{\beta} values in the equation. A simple cone can be created by a \code{parameters} 
vector of c(1.0,0.0,0.0) and is the default if species-specific parameters are not found. The \code{data.frame} has the following members:

\itemize{
   \item species : species code
   \item b0      : \eqn{\beta_0} parameter
   \item b1      : \eqn{\beta_1} parameter
   \item b2      : \eqn{\beta_2} parameter
}
}
\examples{
data(treelist)
library( dplyr )

# use Hann's (1998) largest crown width equations Douglas-fir, western hemlock, bigleaf maple, and red alder
lcw.parms <- data.frame( species=c(202,263,312,351), 
                         b0=c(0.0,         0.0, 0.0,      0.3227140 ),
                         b1=c(0.004363240, 0.0, 0.0,      0.0       ),
                         b2=c(0.6020020,   0.0, 1.470180, 0.0       ))

# estimate distance above crown base: use Hann 1999 for Douglas-fir, Marshall et al. 2003 for western hemlock, and assuming 0.0 for 
# bigleaf maple and red alder
dacb.parms <- data.frame( species=c(202,263,312,351), 
                         d0=c(0.06200, 0.355270, 0.0, 0.0 ) )

# use crown profile parameters for Douglas-fir
cch.parms <- data.frame( species=202, b0=0.929973, b1=-0.1352120, b2=-0.0157579 )

temp <- treelist \%>\% mutate( mcw.hat=mcw( species, dbh ) ) \%>\%
            inner_join( lcw.parms, by="species" ) \%>\%
            inner_join( dacb.parms, by="species" ) \%>\%
            mutate( crown_length = height - htlc,
                    cr = crown_length/height,
                    lcw = mcw.hat * cr^(b0+b1*crown_length+b2*(dbh/height)),
                    dacb = d0 * crown_length )
                    
cch( temp$species, temp$dbh, temp$height,  temp$crown_length, temp$dacb, temp$lcw, temp$tpa, cch.parms )

}
