% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Clark_Evans_R_circle}
\alias{Clark_Evans_R_circle}
\title{Clark_Evans_R_circle() compute Clark Evan's R with Donnelly Edge Correction in a circular plot}
\usage{
Clark_Evans_R_circle(x, y, plotarea, plot_center_x, plot_center_y, plot_radius)
}
\arguments{
\item{x}{: double | vector of x coordinates of trees on plot}

\item{y}{: double | vector of y coordinates of trees on plot}

\item{plotarea}{: double | plot area if polygon not available}

\item{plot_center_x}{: double | x coordinate of plot center.}

\item{plot_center_y}{: double | y coordinate of plot center.}

\item{plot_radius}{: double | radius of circular plot.}
}
\value{
Returns the Clark Evans R statistic.
}
\description{
Compute the Clark and Evans Aggregation Index (R) (1954). The aggregation index R is a measure
of clustering or ordering of trees on a plot. It is the ratio of the observed mean nearest neighbor
distance in the trees to that expected for a Poisson point process of the same intensity. A value 
R > 1 suggests ordering, while R < 1 suggests clustering (unequal inter-tree competition). R has been
proposed as a two-sided, distance-dependent tree competition metric.

This implementation uses Donnelly's correction (Donnelly 1978) for circular plots if a circle is supplied.
If not circle dimensions are not supplied, the function computes `R` with no edge correction and relies on
the supplied `plotarea` value.

\eqn{R = \frac{\frac{\sum{ d_i }}{N}}{(\frac{A}{N})^{0.5}/2}}

where: \eqn{d_i} is the nearest neighbor distance for the ith tree, A is the plot area, and N is the number of 
trees on the plot.
}
\examples{
data(treelistxy)
# TO DO

}
