% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Clark_Evans_Rdc}
\alias{Clark_Evans_Rdc}
\title{Clark_Evans_Rdc() compute Clark Evan's R with Donnelly Edge Correction}
\usage{
Clark_Evans_Rdc(x, y, plot_area, ulx = -1, uly = -1, x_width = 0, y_width = 0)
}
\arguments{
\item{x}{: double | vector of x coordinates of trees on plot}

\item{y}{: double | vector of y coordinates of trees on plot}

\item{plot_area}{: double | area of plot in square units (same units as coordinate system)}

\item{ulx}{: double | upper left corner x of plot rectangle if using Donnelly correction.}

\item{uly}{: double | upper left corner y of plot rectangle if using Donnelly correction.}

\item{x_width}{: double | width of rectangular plot on x axis if using Donnelly correction.}

\item{y_width}{: double | width of rectangular plot on y axis if using Donnelly correction.}
}
\value{
Returns the Clark Evans R statistic.
}
\description{
Compute the Clark and Evans Aggregation Index (R) (1954). The aggregation index R is a measure
of clustering or ordering of trees on a plot. It is the ratio of the observed mean nearest neighbor
distance in the trees to that expected for a Poisson point process of the same intensity. A value 
R > 1 suggests ordering, while R < 1 suggests clustering (unequal inter-tree competition). R has been
proposed as a two-sided, distance-dependent tree competition metric.

This implementation uses Donnelly's correction (Donnelly 1978) for rectangular plots.

\eqn{R = \frac{\frac{\sum{ d_i }}{N}}{(\frac{A}{N})^{0.5}/2}}

where: \eqn{d_i} is the nearest neighbor distance for the ith tree, A is the plot area, and N is the number of 
trees on the plot.
}
\examples{
data(treelistxy)
ulx <- min(treelistxy$x)
uly <- min(treelistxy$y)
x_width <- max(treelistxy$x) - min(treelistxy$x)
y_width <- max(treelistxy$y) - min(treelistxy$y)
clark_evans_Rdc( treelistxy$x, treelistxy$y, x_width*y_width, ulx, uly, x_width, y_width )


}
