% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{curtis_rd}
\alias{curtis_rd}
\title{curtis_rd() compute Curtis' Relative Density}
\usage{
curtis_rd(dbh, expansion, imperial_units = TRUE)
}
\arguments{
\item{dbh}{: double | vector of diameters at breast height}

\item{expansion}{: double | vector of expansion factors}

\item{imperial_units}{: bool   | TRUE = imperial (default), FALSE = metric}
}
\value{
Returns the Curtis' Relative Density.
}
\description{
Compute Curtis' Relative Density (Curtis 1982).

The expression \eqn{RD = G/(Dg^{0.5})}, where G is basal area and Dg is quadratic mean
stand diameter, provides a simple and convenient scale of relative stand density for Douglas-fir, 
equivalent to other generally accepted diameter-based stand density measures.
}
\examples{
data(treelist)
curtis_rd( treelist$dbh, treelist$tpa )

}
