% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dominant_dbh}
\alias{dominant_dbh}
\title{dominant_dbh() compute the average dbh of trees in the dominant cohort.}
\usage{
dominant_dbh(height, dbh, expansion, dominant_cohort_size, method = 0L)
}
\arguments{
\item{height}{: double | vector of total height}

\item{dbh}{: double | vector of diameter at breast height}

\item{expansion}{: double | vector of expansion factors}

\item{dominant_cohort_size}{: int    | number of trees in the dominant height cohort}

\item{method}{: int    | 0 (default), 1, or 2 (see below for definitions)}
}
\value{
Returns the specified dominant dbh.
}
\description{
Compute the expansion factor weighted average dbh of trees in the defined dominant tree cohort. Each method
defines the cohort differently:

\itemize{
   \item 0 : average dbh of the dominant_cohort_size trees by decreasing dbh
   \item 1 : average dbh of the dominant_cohort_size trees by decreasing height
   \item 2 : Lorey dbh (quadratic mean diameter)
}

The cohort size is ignored and should be 0.0 for Lorey dbh (option 2).
}
\examples{
data(treelist)
# compute the dbh of the 40 largest trees by dbh
dominant_dbh( treelist$height, treelist$dbh, treelist$tpa, 40, 0 )
# compute the dbh of the 100 largest trees by height
dominant_dbh( treelist$height, treelist$dbh, treelist$tpa, 100, 1 )
# compute the Lorey dbh 
dominant_dbh( treelist$height, treelist$dbh, treelist$tpa, 0, 2 )

}
