% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dominant_height}
\alias{dominant_height}
\title{dominant_height() compute the average height of trees in the dominant cohort.}
\usage{
dominant_height(height, dbh, expansion, dominant_cohort_size, method = 0L)
}
\arguments{
\item{height}{: double | vector of total height}

\item{dbh}{: double | vector of diameter at breast height}

\item{expansion}{: double | vector of expansion factors}

\item{dominant_cohort_size}{: int    | number of trees in the dominant height cohort}

\item{method}{: int    | 0 (default), 1, or 2 (see below for definitions)}
}
\value{
Returns the specified dominant height.
}
\description{
Compute the expansion factor weighted average height of trees in the defined dominant tree cohort. Each method
defines the cohort differently:

\itemize{
   \item 0 : average height of the dominant_cohort_size trees by decreasing dbh
   \item 1 : average height of the dominant_cohort_size trees by decreasing height
   \item 2 : Lorey height (height of the tree of average basal area)
}
}
