% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Glover_Hool}
\alias{Glover_Hool}
\title{Glover_Hool() compute the Glover and Hool competition index for each tree.}
\usage{
Glover_Hool(dbh, expansion, use_arithmetic = TRUE, imperial_units = TRUE)
}
\arguments{
\item{dbh}{: double | vector of diameter at breast height}

\item{expansion}{: double | vector of expansion factors}

\item{use_arithmetic}{: bool   | TRUE = use arithmetic mean (default), FALSE = use quadratic mean}

\item{imperial_units}{: bool   | TRUE = imperial (default), FALSE = metric}
}
\value{
Returns a vector of competition indicies for each tree in their original order.
}
\description{
Compute the Glover and Hool (1979) competition index. The index is interpreted as the ratio
of a tree's basal area to the basal area of the tree of mean diameter. Glover and Hool used
the arithmetic mean and a common variation is to use the quadratic mean (use \code{use_arithmetic}
flag to select the desired method).

The index \eqn{G_i} is:

\eqn{G_i = dbh_i^2 / \overline{dbh}^2}

where: \eqn{dbh_i} is the diameter of tree \code{i} and \eqn{\overline{dbh}} is the mean diameter (either arithmetic or quadratic).
}
\examples{
data(treelist)
# compute the Glover and Hool index with each mean
Glover_Hool( treelist$dbh, treelist$tpa, use_arithmetic=T )
Glover_Hool( treelist$dbh, treelist$tpa, use_arithmetic=F )

}
