% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hd_fit}
\alias{hd_fit}
\title{hd_fit() Fit height-dbh curve}
\usage{
hd_fit(dbh, height, bh = 4.5)
}
\arguments{
\item{dbh}{: double | vector of diameter at breast height}

\item{height}{: double | vector of total heights}

\item{bh}{: double | height to breast height (default is 4.5 feet; use 1.37 for metric measurements)}
}
\value{
A vector of estimated parameters suitable for use in \code{hd_predict()}.
}
\description{
Fits a height-diameter curve to provided \code{height} and \code{dbh} vectors. The functional form is:

\eqn{\widehat{height} = BH + e^{(\beta_0 + \beta_1 dbh^{-\beta_2})}}

where BH = height to breast height (e.g., 4.5 feet for imperial measurements, 1.37 meters for metric), and \eqn{\beta} s are
parameters to be estimated. Fitting height-dbh curves is often difficult due to limited measurement data (either in observation
count or in range, or both). We are using David Marshall's technique of fitting a linearized form of the equation while iterating
over a range of \eqn{\beta_2} values (-0.1 to -1.0). The \eqn{\beta_2} value yielding the lowest sum of squared errors (SSE) is chosen.
}
\examples{
data(treelist )
hd.model <- hd_fit( treelist$dbh, treelist$height )
plot( treelist$dbh, hd_predict( hd.model, treelist$dbh ), col="green" )
points( treelist$dbh, treelist$height )

}
