% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mcw}
\alias{mcw}
\title{mcw() Estimate maximum crown width for a species}
\usage{
mcw(fia, dbh, imperial_units = TRUE, default_fia = 202L)
}
\arguments{
\item{fia}{: int    | vector of FIA species code}

\item{dbh}{: double | vector of diameter at breast height}

\item{imperial_units}{: bool   | TRUE = imperial (default), FALSE = metric}

\item{default_fia}{: int    | FIA species code to use if supplied FIA code does not have parameters (default = 202 (Douglas-fir))}
}
\value{
Returns numeric vector with maximum crown width in the original tree order (unsorted).
}
\description{
Estimate the maximum (open-grown) crown width for trees using publically available equation parameters. A list of available species
can be found using \code{mcw_species()}. If an FIA species code does not have a parameter set, the \code{default_fia} species is used.
}
\examples{
# compute maximum crown width for a 10 inch Douglas-fir, red alder, and Atlantic White-cedar (parameters not available, will default to Douglas-fir)
mcw( c(202,351,43), c(10,10,10) )

}
