% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mcw_species}
\alias{mcw_species}
\title{mcw_species() List species with available Maximum Crown Width equation parameters}
\usage{
mcw_species()
}
\value{
A data.frame with:
\itemize{
   \item fia : FIA species code
   \item name : common name
}
}
\description{
Builds a data.frame with the FIA species codes and common names of species with maximum crown
width equation parameters available for use.
}
\examples{
# list available maximum crown width species
mcw_species()

}
