% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reineke_sdi}
\alias{reineke_sdi}
\title{reineke_sdi() compute Reineke's Stand Density Index}
\usage{
reineke_sdi(dbh, expansion, imperial_units = TRUE)
}
\arguments{
\item{dbh}{: double | vector of diameters at breast height}

\item{expansion}{: double | vector of expansion factors}

\item{imperial_units}{: bool   | TRUE = imperial (default), FALSE = metric}
}
\value{
Returns the Reineke's Stand Density Index.
}
\description{
Compute Reineke's Stand Density Index \url{https://research.fs.usda.gov/treesearch/60134}.

Reineke (1933) developed a stand density index (SDI) that relates the current
stand density to an equivalent density in a stand with a quadratic mean diameter (Dq) of 10 inches.
Reineke’s SDI can be expressed as:
\eqn{SDI = N(Dq/10)^b}
where SDI = Reineke’s stand-density index, N = trees per acre, Dq = quadratic mean diameter (inches),
b = exponent of Reineke’s equation, often reported to equal –1.605.

The function allows for computing the equivalent index in metric units.
}
