% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{relative_spacing}
\alias{relative_spacing}
\title{relative_spacing() compute Wilson's Relative Spacing}
\usage{
relative_spacing(expansion, dominant_height, imperial = TRUE)
}
\arguments{
\item{expansion}{: double | vector of expansion factors}

\item{dominant_height}{: double | dominant height of the stand}

\item{imperial_units}{: bool   | TRUE = imperial (default), FALSE = metric}
}
\value{
Returns the relative spacing as a fraction of dominant height.
}
\description{
Compute Wilson's Relative Spacing (average tree spacing relative to dominant height).
}
