% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ccf}
\alias{ccf}
\title{ccf() compute crown competition factor}
\usage{
ccf(crown_width, expansion, imperial_units = TRUE)
}
\arguments{
\item{crown_width}{: double | vector of open-grown crown widths}

\item{expansion}{: double | vector of expansion factors}

\item{imperial_units}{: bool   | TRUE = imperial (default), FALSE = metric}
}
\value{
Returns the crown competition factor.
}
\description{
Compute Crown Competition Factor 
\url{https://cmapspublic3.ihmc.us/rid=1N4TSFQX6-GWW4BN-14PZ/Crown\%20competition\%20-\%20A\%20measure\%20of\%20density.pdf}.
Krajicek, J.E., K.A. Brinkman, and F.S. Gingrich.  1961.  Crown competition: a measure of density.  For. Sci. 7:36 – 42.  

Crown competition factor is the ratio of the open-grown crown area of all trees as a percentage of an acre:
\eqn{CCF = 100 \sum{CA}/43560}, 
where CA is the open-grown crown area of a given tree.

The function allows for computing the equivalent index in metric units.
}
