% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ccfl}
\alias{ccfl}
\title{ccfl() compute crown competition factor in larger trees.}
\usage{
ccfl(dbh, mcw, expansion, imperial_units = TRUE)
}
\arguments{
\item{dbh}{: double | vector of diameter at breast height}

\item{mcw}{: double | vector of maximum crown width for the tree record}

\item{expansion}{: double | vector of expansion factors}

\item{imperial_units}{: bool   | TRUE = imperial, FALSE = metric}
}
\value{
Returns numeric vector with the crown competition factor in larger trees in the original tree order (unsorted):
}
\description{
Sorts the dbh, mcw, and expansion vectors in decreasing order of dbh and computes cumulative maximum crown area in larger trees. Ties in dbh are
handled appropriately (each tree has the same ccfl).

The algorithm computes the percentage of an acre or hectare (depending on \code{imperial_units}) covered by each tree
record's crown.  It uses the maximum crown width to estimate the coverage for each tree and expands it to a per
acre or hectare basis using the supplied expansion factor.
}
\examples{
library( dplyr )
data(treelist)
ccfl( treelist$dbh, mcw( treelist$species, treelist$dbh ), treelist$tpa )

}
