% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cch}
\alias{cch}
\title{cch() compute closure at tree tip without using interpolation.}
\usage{
cch(
  dbh,
  height,
  crown_length,
  dacb,
  lcw,
  expansion,
  parameters,
  imperial_units = TRUE
)
}
\arguments{
\item{dbh}{: double | vector of diameter at breast height}

\item{height}{: double | vector of total height}

\item{crown_length}{: double | vector of crown lengths}

\item{dacb}{: double | vector of distance above crown base to largest width for each tree}

\item{lcw}{: double | vector of largest crown width for each tree}

\item{expansion}{: double | vector of expansion factors}

\item{parameters}{: double | vector of parameters (see below)}

\item{imperial_units}{: bool   | TRUE = imperial (default), FALSE = metric}
}
\value{
A vector of \code{cch} values for each tree.
}
\description{
Computes the crown area as a fraction of an acre or hectare (depending on \code{imperial_units}) that lies in a horizontal
plane tangential to the tip of the tree.

cch() depends on \code{dacb} and \code{lcw} which must be estimated for each tree (normally species and size dependent)
independently. 

The \code{parameters} vector contains coefficients for an equation that adjusts \code{lcw} for the position within the crown of the form:
\eqn{adjustment = rp^{(\beta_0 + \beta_1 rp^{0.5} + \beta_2 height / dbh)}}
where \code{rp} is the relative position in the crown, and \code{adjustment} is the estimated percentage of the \code{lcw} at that point
in the crown. The \code{parameters} are the \eqn{\beta} values in the equation. A simple cone can be created by a \code{parameters} 
vector of c(1.0,0.0,0.0).
}
