% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hd_predict}
\alias{hd_predict}
\title{hd_predict() Use parameter estimates from \code{hd_fit(0)} to predict heights for a vector of \code{dbh}.}
\usage{
hd_predict(hd_parameters, dbh, bh = 4.5)
}
\arguments{
\item{dbh}{: double | vector of diameter at breast height}

\item{bh}{: double | height to breast height (default is 4.5 feet; use 1.37 for metric measurements)}

\item{parms}{: double | vector of height-dbh equation parameters}
}
\value{
A vector of predicted heights for each \code{dbh} supplied.
}
\description{
Predicts heights for a \code{dbh} vector. See \link{hd_fit} for details on the parameter estimates used in the prediction.
}
\examples{
data(treelist )
hd.model <- hd_fit( treelist$dbh, treelist$height )
plot( treelist$dbh, hd_predict( hd.model, treelist$dbh ), col="green" )
points( treelist$dbh, treelist$height )

}
