% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Hegyi}
\alias{Hegyi}
\title{Hegyi() compute Hegyi's distance weighted size ratio.}
\usage{
Hegyi(x, y, dbh, poly_x, poly_y, imperial_units = FALSE)
}
\arguments{
\item{x}{: double | vector of x coordinates of trees on plot}

\item{y}{: double | vector of y coordinates of trees on plot}

\item{dbh}{: double | vector of diameter at breast height}

\item{poly_x}{: double | vector of plot polygon x coordinates (use 0 if no plot boundaries available).}

\item{poly_y}{: double | vector of plot polygon y coordinates (use 0 if no plot boundaries available).}

\item{imperial_units}{: bool   | TRUE = imperial, FALSE = metric (default)}
}
\value{
Returns the Hegyi's distance weighted size ratio for each tree. A `NaN` is returned if a tree has a `dbh` of 0.0.
}
\description{
Compute Hegyi's (1974) distance-weighted size ratio (a two-sided competition index) for trees within a 6-meter fixed radius plot.
The ratio for the ith tree in the 6-meter radius plot is:

\eqn{heygi_i = \sum{\frac{ba_j/ba_i}{d_{ij}}}}

where \eqn{ba_j} and \eqn{ba_i} are the basal areas of the jth and ith tree respectively, \eqn{d_{ij}} is the distance between tree i and j.

Trees from a plot of arbitrary size can be used. The Hegyi ratio for each tree will be computed based on its neighbors within the 6-meter boundary. 
If \code{imperial_units} is TRUE, the coordinates will be converted to meters prior to calculations.

`Hegyi` adjusts for edge effects using Ripley's (1977) edge correction if plot boundaries are supplied, otherwise edge effects are ignored.
}
\examples{
data(treelistxy)
min_x <- min(treelistxy$x)
min_y <- min(treelistxy$y)
max_x <- max(treelistxy$x)
max_y <- max(treelistxy$y)
poly_x <- c(min_x, max_x, max_x, min_x)
poly_y <- c(min_y, min_y, max_y, max_y)
Hegyi( treelistxy$x, treelistxy$y, treelistxy$dbh, poly_x, poly_y, imperial_units=T )

}
